StimStudy/requirements.txt
python-dotenv==1.1.0
google-generativeai==0.3.2
pyttsx3>=2.90
moviepy>=1.0.3
Pillow>=8.0.0
selenium>=4.1.0
elevenlabs==0.1.0
fish-audio-sdk==2025.4.2
mutagen>=1.46.0
tqdm>=4.65.0
